@@ -1,10 +1,6 @@
 {
   "name": "treehillstudio/pdfresource",
   "description": "Convert MODX resources to PDF files using mPDF",
-  "require": {
-    "mpdf/mpdf": "^6.0"
-  },
-  "license": "GPL-2.0-or-later",
   "authors": [
     {
       "name": "Thomas Jakobi",
@@ -12,9 +8,14 @@
       "homepage": "https://github.com/Jako"
     }
   ],
-  "config": {
-    "platform": {
-      "php": "5.4"
+  "license": "GPL-2.0-or-later",
+  "require": {
+    "php": ">=7.2.5",
+    "mpdf/mpdf": "^8.0"
+  },
+  "autoload": {
+    "psr-4": {
+      "TreehillStudio\\PDFResource\\": "src/"
     }
   }
 }
