@@ -9,21 +9,24 @@
  * @var array $scriptProperties
  */
 
-$className = 'PDFResource' . $modx->event->name;
+$className = 'TreehillStudio\PDFResource\Plugins\Events\\' . $modx->event->name;
 
 $corePath = $modx->getOption('pdfresource.core_path', null, $modx->getOption('core_path') . 'components/pdfresource/');
 /** @var PDFResource $pdfresource */
-$pdfresource = $modx->getService('pdfresource', 'PDFResource', $corePath . 'model/pdfresource/', array(
+$pdfresource = $modx->getService('pdfresource', 'PDFResource', $corePath . 'model/pdfresource/', [
     'core_path' => $corePath
-));
+]);
 
 if ($pdfresource) {
-    $modx->loadClass('PDFResourcePlugin', $pdfresource->getOption('modelPath') . 'pdfresource/events/', true, true);
-    $modx->loadClass($className, $pdfresource->getOption('modelPath') . 'pdfresource/events/', true, true);
     if (class_exists($className)) {
-        /** @var PDFResourcePlugin $handler */
         $handler = new $className($modx, $scriptProperties);
-        $handler->run();
+        if (get_class($handler) == $className) {
+            $handler->run();
+        } else {
+            $modx->log(xPDO::LOG_LEVEL_ERROR, $className. ' could not be initialized!', '', 'PDFResource Plugin');
+        }
+    } else {
+        $modx->log(xPDO::LOG_LEVEL_ERROR, $className. ' was not found!', '', 'PDFResource Plugin');
     }
 }
 
